+++ csvtk/cmd/cut.go
@@ -120,7 +120,7 @@ Examples:
 
 		parseHeaderRow := needParseHeaderRow // parsing header row
 		handleHeaderRow := len(colnames) > 0
-		var colnames2fileds map[string]int // column name -> field
+		var colnames2fileds map[string][]int // column name -> []field
 		var colnamesMap map[string]*regexp.Regexp
 
 		checkFields := true
@@ -142,12 +142,13 @@ Examples:
 				if parseHeaderRow { // parsing header row
 					if len(fields) == 0 { // user gives the colnames
 						// colnames
-						colnames2fileds = make(map[string]int, len(record))
+						colnames2fileds = make(map[string][]int, len(record))
 						for i, col := range record {
-							if ignoreCase {
-								col = strings.ToLower(col)
+							if _, ok := colnames2fileds[col]; !ok {
+								colnames2fileds[col] = []int{i + 1}
+							} else {
+								colnames2fileds[col] = append(colnames2fileds[col], i+1)
 							}
-							colnames2fileds[col] = i + 1
 						}
 
 						// colnames from user
@@ -197,7 +198,7 @@ Examples:
 								}
 
 								if !ok {
-									fields = append(fields, colnames2fileds[col])
+									fields = append(fields, colnames2fileds[col]...)
 								}
 							}
 						} else {
@@ -210,21 +211,22 @@ Examples:
 								for _, col := range colnames {
 									for _, col2 := range record {
 										if colnamesMap[col].MatchString(col2) {
-											i = colnames2fileds[col2]
-											if uniqColumn {
-												if _, ok = flags[i]; !ok {
+											for _, i = range colnames2fileds[col2] {
+												if uniqColumn {
+													if _, ok = flags[i]; !ok {
+														fields = append(fields, i)
+														flags[i] = struct{}{}
+													}
+												} else {
 													fields = append(fields, i)
-													flags[i] = struct{}{}
 												}
-											} else {
-												fields = append(fields, i)
 											}
 										}
 									}
 								}
 							} else {
 								for _, col := range colnames {
-									fields = append(fields, colnames2fileds[col])
+									fields = append(fields, colnames2fileds[col]...)
 								}
 							}
 						}