+++ csvtk/cmd/freq.go
@@ -127,7 +127,7 @@ var freqCmd = &cobra.Command{
 
 		parseHeaderRow := needParseHeaderRow // parsing header row
 		printHeaderRow := needParseHeaderRow
-		var colnames2fileds map[string]int // column name -> field
+		var colnames2fileds map[string][]int // column name -> []field
 		var colnamesMap map[string]*regexp.Regexp
 
 		checkFields := true
@@ -147,9 +147,14 @@ var freqCmd = &cobra.Command{
 			for _, record := range chunk.Data {
 				N++
 				if parseHeaderRow { // parsing header row
-					colnames2fileds = make(map[string]int, len(record))
+					colnames2fileds = make(map[string][]int, len(record))
 					for i, col := range record {
-						colnames2fileds[col] = i + 1
+						if _, ok := colnames2fileds[col]; !ok {
+							colnames2fileds[col] = []int{i + 1}
+						} else {
+							// checkError(fmt.Errorf("duplicate colnames not allowed: %s", col))
+							colnames2fileds[col] = append(colnames2fileds[col], i+1)
+						}
 					}
 					colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 					i := 0
@@ -189,7 +194,7 @@ var freqCmd = &cobra.Command{
 								_, ok = colnamesMap[col]
 							}
 							if ok {
-								fields = append(fields, colnames2fileds[col])
+								fields = append(fields, colnames2fileds[col]...)
 							}
 						}
 					}