+++ csvtk/cmd/inter.go
@@ -110,7 +110,7 @@ Attention:
 			csvReader.Run()
 
 			parseHeaderRow := needParseHeaderRow // parsing header row
-			var colnames2fileds map[string]int   // column name -> field
+			var colnames2fileds map[string][]int // column name -> []field
 			var colnamesMap map[string]*regexp.Regexp
 			var HeaderRow []string
 
@@ -129,9 +129,14 @@ Attention:
 
 				for _, record := range chunk.Data {
 					if parseHeaderRow { // parsing header row
-						colnames2fileds = make(map[string]int, len(record))
+						colnames2fileds = make(map[string][]int, len(record))
 						for i, col := range record {
-							colnames2fileds[col] = i + 1
+							if _, ok := colnames2fileds[col]; !ok {
+								colnames2fileds[col] = []int{i + 1}
+							} else {
+								checkError(fmt.Errorf("duplicate colnames not allowed: %s", col))
+								colnames2fileds[col] = append(colnames2fileds[col], i+1)
+							}
 						}
 						colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 						for _, col := range colnames {
@@ -168,7 +173,7 @@ Attention:
 									_, ok = colnamesMap[col]
 								}
 								if ok {
-									fields = append(fields, colnames2fileds[col])
+									fields = append(fields, colnames2fileds[col]...)
 								}
 							}
 						}