+++ csvtk/cmd/round.go
@@ -105,7 +105,7 @@ var roundCmd = &cobra.Command{
 
 			parseHeaderRow := needParseHeaderRow // parsing header row
 			parseHeaderRow2 := needParseHeaderRow
-			var colnames2fileds map[string]int // column name -> field
+			var colnames2fileds map[string][]int // column name -> []field
 			var colnamesMap map[string]*regexp.Regexp
 
 			checkFields := true
@@ -126,9 +126,13 @@ var roundCmd = &cobra.Command{
 
 				for _, record := range chunk.Data {
 					if parseHeaderRow { // parsing header row
-						colnames2fileds = make(map[string]int, len(record))
+						colnames2fileds = make(map[string][]int, len(record))
 						for i, col := range record {
-							colnames2fileds[col] = i + 1
+							if _, ok := colnames2fileds[col]; !ok {
+								colnames2fileds[col] = []int{i + 1}
+							} else {
+								colnames2fileds[col] = append(colnames2fileds[col], i+1)
+							}
 						}
 						colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 						for _, col := range colnames {
@@ -165,7 +169,7 @@ var roundCmd = &cobra.Command{
 									_, ok = colnamesMap[col]
 								}
 								if ok {
-									fields = append(fields, colnames2fileds[col])
+									fields = append(fields, colnames2fileds[col]...)
 								}
 							}
 						}