+++ csvtk/cmd/split.go
@@ -114,7 +114,7 @@ Note:
 
 		parseHeaderRow := needParseHeaderRow // parsing header row
 		printHeaderRow := needParseHeaderRow
-		var colnames2fileds map[string]int // column name -> field
+		var colnames2fileds map[string][]int // column name -> []field
 		var colnamesMap map[string]*regexp.Regexp
 
 		checkFields := true
@@ -131,9 +131,14 @@ Note:
 
 			for _, record := range chunk.Data {
 				if parseHeaderRow { // parsing header row
-					colnames2fileds = make(map[string]int, len(record))
+					colnames2fileds = make(map[string][]int, len(record))
 					for i, col := range record {
-						colnames2fileds[col] = i + 1
+						if _, ok := colnames2fileds[col]; !ok {
+							colnames2fileds[col] = []int{i + 1}
+						} else {
+							checkError(fmt.Errorf("duplicate colnames not allowed: %s", col))
+							colnames2fileds[col] = append(colnames2fileds[col], i+1)
+						}
 					}
 					colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 					for _, col := range colnames {
@@ -169,7 +174,7 @@ Note:
 								_, ok = colnamesMap[col]
 							}
 							if ok {
-								fields = append(fields, colnames2fileds[col])
+								fields = append(fields, colnames2fileds[col]...)
 							}
 						}
 					}