+++ csvtk/cmd/splitxlsx.go
@@ -126,7 +126,7 @@ Weakness : Complicated sheet structures are not well supported, e.g.,
 
 		parseHeaderRow := needParseHeaderRow // parsing header row
 		printHeaderRow := needParseHeaderRow
-		var colnames2fileds map[string]int // column name -> field
+		var colnames2fileds map[string][]int // column name -> []field
 		var colnamesMap map[string]*regexp.Regexp
 
 		checkFields := true
@@ -141,9 +141,14 @@ Weakness : Complicated sheet structures are not well supported, e.g.,
 		checkError(err)
 		for rowIndex, record := range rows {
 			if parseHeaderRow { // parsing header row
-				colnames2fileds = make(map[string]int, len(record))
+				colnames2fileds = make(map[string][]int, len(record))
 				for i, col := range record {
-					colnames2fileds[col] = i + 1
+					if _, ok := colnames2fileds[col]; !ok {
+						colnames2fileds[col] = []int{i + 1}
+						checkError(fmt.Errorf("duplicate colnames not allowed: %s", col))
+					} else {
+						colnames2fileds[col] = append(colnames2fileds[col], i+1)
+					}
 				}
 				colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 				for _, col := range colnames {
@@ -179,7 +184,7 @@ Weakness : Complicated sheet structures are not well supported, e.g.,
 							_, ok = colnamesMap[col]
 						}
 						if ok {
-							fields = append(fields, colnames2fileds[col])
+							fields = append(fields, colnames2fileds[col]...)
 						}
 					}
 				}