+++ csvtk/cmd/summary.go
@@ -220,7 +220,7 @@ Available operations:
 		csvReader.Run()
 
 		parseHeaderRow := needParseHeaderRow // parsing header row
-		var colnames2fileds map[string]int   // column name -> field
+		var colnames2fileds map[string][]int // column name -> []field
 		var colnamesMap map[string]*regexp.Regexp
 		var HeaderRow []string
 		var isHeaderRow bool
@@ -245,9 +245,14 @@ Available operations:
 
 			for _, record := range chunk.Data {
 				if parseHeaderRow { // parsing header row
-					colnames2fileds = make(map[string]int, len(record))
+					colnames2fileds = make(map[string][]int, len(record))
 					for i, col := range record {
-						colnames2fileds[col] = i + 1
+						if _, ok := colnames2fileds[col]; !ok {
+							colnames2fileds[col] = []int{i + 1}
+						} else {
+							checkError(fmt.Errorf("duplicate colnames not allowed: %s", col))
+							colnames2fileds[col] = append(colnames2fileds[col], i+1)
+						}
 					}
 					colnamesMap = make(map[string]*regexp.Regexp, len(colnames))
 					for _, col := range colnames {
@@ -285,13 +290,13 @@ Available operations:
 							}
 							if (negativeFields && !ok) || (!negativeFields && ok) {
 								if _, ok = fieldsStrsDMap[col]; ok {
-									fieldsD = append(fieldsD, colnames2fileds[col])
-									statsI[colnames2fileds[col]] = stats[col]
+									fieldsD = append(fieldsD, colnames2fileds[col][0])
+									statsI[colnames2fileds[col][0]] = stats[col]
 								}
 								if _, ok = fieldsStrsGMap[col]; ok {
-									fieldsG = append(fieldsG, colnames2fileds[col])
+									fieldsG = append(fieldsG, colnames2fileds[col][0])
 								}
-								fields = append(fields, colnames2fileds[col])
+								fields = append(fields, colnames2fileds[col][0])
 							}
 						}
 					}