+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - [csvtk v0.26.0](https://github.com/shenwei356/csvtk/releases/tag/v0.26.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.26.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.26.0)
   - `csvtk`: near all commands skip empty file now. [#204](https://github.com/shenwei356/csvtk/issues/204)
+  - `csvtk cut/fmtdate/freq/grep/rename/rename2/replace/round`: allow duplicated column names.
   - `csvtk xlsx2csv`: fix bug wheree `xlsx2csv` treats small number (padj < 1e-25) as 0. It's solved by updating the excelize package. [#261](https://github.com/shenwei356/csvtk/issues/201)
   - `csvtk join`: new flag for adding filename as column name prefix. by @tetedange13 [#202](https://github.com/shenwei356/csvtk/issues/202)
 - [csvtk v0.25.0](https://github.com/shenwei356/csvtk/releases/tag/v0.25.0)